docker-specification/producer/requirements.txt
praw
kafka-python
pandas